refactor(Exercise): move sources part into the main content
